# main/docs/swagger.yaml
basePath: /api/
info:
  contact: {}
  description: 运维平台
  title: 运维平台
  version: 1.0.0
paths:
  /addAppDetail/:
    post:
      description: 添加应用信息
      operationId: add-application-detail
      parameters:
      - description: app名字
        in: formData
        name: app_name
        required: true
        type: string
      - description: 主机名
        in: formData
        name: host
        required: true
        type: string
      - description: 目录
        in: formData
        name: dir
        required: true
        type: string
      - description: 类型
        in: formData
        name: type
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
...
  /addServer/:
    post:
      description: 添加资产信息
      operationId: add-server
      parameters:
      - description: ip
        in: formData
        name: ip
        required: true
        type: string
      - description: idc名字
        in: formData
        name: idc
        type: string
      - description: 备注
        in: formData
        name: comment
        type: string
      - description: 用途
        in: formData
        name: use
        type: string
      - description: 机柜号
        in: formData
        name: cabinet
        type: string
      - description: U位
        in: formData
        name: uPosition
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已开始创建"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
...
  /appDetail/{id}/:
    delete:
      description: 删除单条信息
      operationId: delete-application-detail
      parameters:
      - description: 应用id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
...
  /deleteAllAppDetail/:
    delete:
      description: 批量删除应用信息
      operationId: delete-all-application-detail
      parameters:
      - in: formData
        items:
          type: integer
        name: detail_ids
        type: array
      responses:
        "200":
          description: '{"code":200,"msg":"已全部删除"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
...
  /deleteAllServer/:
    delete:
      description: 批量删除资产信息
      operationId: delete-all-server
      parameters:
      - in: formData
        items:
          type: integer
        name: server_ids
        type: array
      responses:
        "200":
          description: '{"code":200,"msg":"已全部删除"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
  /getAllAppDetail/:
    get:
      description: 获取全部应用信息
      operationId: get-all-application-detail
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":[...]}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
  /getAllServer/:
    get:
      description: 获取全部资产信息
      operationId: get-all-server
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":[...]}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
...
  /operationApp/:
    post:
      description: 操作应用
      operationId: operation-application
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: string
      - description: 操作动作
        in: formData
        name: action
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已开始操作"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
  /server/{id}/:
    delete:
      description: 删除资产信息
      operationId: delete-server
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
    get:
      description: 获取单台资产信息
      operationId: get-server
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":{object}}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
    patch:
      description: 更新资产信息
      operationId: update-server
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: idc名字
        in: formData
        name: idc
        type: string
      - description: 备注
        in: formData
        name: comment
        type: string
      - description: 用途
        in: formData
        name: use
        type: string
      - description: 机柜号
        in: formData
        name: cabinet
        type: string
      - description: U位
        in: formData
        name: uPosition
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - asset
...
  /updateAppDetailState/:
    put:
      description: 更新应用状态
      operationId: update-application-detail
      parameters:
      - description: app名字
        in: formData
        name: app_name
        required: true
        type: string
      - description: 主机名
        in: formData
        name: host
        required: true
        type: string
      - description: 目录
        in: formData
        name: dir
        required: true
        type: string
      - description: 类型
        in: formData
        name: type
        required: true
        type: string
      - description: 状态
        in: formData
        name: state
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已更新"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Application
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"